rework unsafe handling of SIGINT/SIGTERM
Signals are asynchronous. It is unsafe to perform such complex
operations in a signal handler. I changed signal handler
to just set a flag, and added an event source which returns an event
when this variable is set. The action is to stop event loop.
Execution then falls through to program exit.

Signed-off-by: Denys Vlasenko <vda.linux@googlemail.com>