lib/Plugins/CCpp.cpp: move functions out of class where appropriate.
I was struggling to read the code where classes have member functions
with no apparent reasons to be such: they do not use any members
of the class.

Having them members of the class have only disadvantages:
they need to be declared in the class (thus you need to touch TWO files),
they cannot be shared among different classes,
they look confusing at callsites - the code falsely suggests
that function uses or alters object's state,
they also are a bit less efficient.

I was hesitating to do so before, but now I was told that
there is a plan to gradually move to C implementation...
then this change makes it a bit closer to that.

Signed-off-by: Denys Vlasenko <vda.linux@googlemail.com>