abrt-gui: move dbus signal handlers init where they should be
Old code was installing dbus signal handlers as if they are
somehow related to the remote object we work with, but
this is wrong - in reality, dbus signal handlers are installed
on the dbus connection, they are not related to the object.
Now code better reflect this.
Also, remove AnalyzeComplete dbus handler - there is not such dbus signal.

Run tested.

Signed-off-by: Denys Vlasenko <vda.linux@googlemail.com>