This patch makes abrtd spawn a new process, abrt-server, for every socket client.
This allows for simpler timeout handling
using SIGALRM, and makes timers for tracking client timeouts
unnecessary.

This also allows for debugging and regression testing
of abrt-server separately - it can be simply run from command-line
and fed commands on stdin.
Also it provides a better fault isolation - crash in
abrt-server does not take down abrtd.

The code is based on dumpsocket.{cpp,h}. Most of dumpsocket.cpp
goes to abrt-server.c, a small part goes to Daemon.cpp (i.e. to abrtd).

This change will not compile - it does not have dumpsocket.cpp
renamed to abrt-server.c, which makes the change easier to read.

The next change, which I'll commit after this one,
contains only the renaming.

Signed-off-by: Denys Vlasenko <dvlasenk@redhat.com>