factor out headers so that "report lib" API is separated out of abrt API
This patch minimally affects code per se - it adds
create_crash_dump_dir() function which takes in-memory
representation of the dump (map_crash_data_t object)
and creates an on-disk representation.
Then it returns pointer to struct dump_dir.

With this function, it is possible to run an event
on a user-constructed map_crash_data_t:

    map_crash_data_t cd;
    add_to_crash_data(cd, "foo", "bar");

    struct dump_dir *dd = create_crash_dump_dir(cd);
    char *dir_name = strdup(dd->dd_dir);
    dd_close(dd);
    run_event(run_state, dir_name, event);
    delete_crash_dump_dir(dir_name);

which is, basically, what report library is about.

The biggest part of the patch is reshuffling of
header files, with the following result: three header
files which are not cluttered by other ABRT stuff,
and expose the following API each:

crash_dump.h - in-memory crash dump data structs and ops
dump_dir.h   - on-disk crash dump data structs and ops
run_event.h  - run_event() and friends

These is a test application, test_report.cpp.cpp,
which demonstrates (and tests) usage of these headers.
(grep for "test-report" and enable it in build system
if you want it to be built).
It creates a temporary dump in
/var/run/abrt/tmp-NNNN-NNNNN and runs "report" event.
Deleting of temp dump is disabled for testing purposes -
you might want to see the contents.

Here is how the binary looks like:

   text    data     bss     dec     hex filename
   3730     668      48    4446    115e /usr/bin/test-report
        linux-vdso.so.1 =>  (0x00007ffffa80f000)
        libabrt.so.0 => /usr/lib64/libabrt.so.0 (0x00007fad1f35e000)
        libglib-2.0.so.0 => /lib64/libglib-2.0.so.0 (0x0000003c1f200000)
        libstdc++.so.6 => /usr/lib64/libstdc++.so.6 (0x0000003c2c200000)
        libm.so.6 => /lib64/libm.so.6 (0x00007fad1f0da000)
        libgcc_s.so.1 => /lib64/libgcc_s.so.1 (0x0000003c28e00000)
        libc.so.6 => /lib64/libc.so.6 (0x00007fad1ed5b000)
        /lib64/ld-linux-x86-64.so.2 (0x00007fad1f570000)

Next step would be to clean up the namespace,
then rename libabrt.so into libreport.so,
then split it into a separate package so that we can install it
without installing other two abrt .so files.

Signed-off-by: Denys Vlasenko <dvlasenk@redhat.com>