use run_state->children_count == 0 check for "event not conf'd" condition
run_event_on_FOO() was returning -1 when it found that not even one
program was run on the requested event. Which is not a very natural
return value: in many cases this isn't an error.

This change makes it return 0 in this case. It is ok since now we have
run_state->children_count member which can be checked for 0 by those
callers which want to detect this condition.

Signed-off-by: Denys Vlasenko <dvlasenk@redhat.com>