show more useful columns in abrt-gtk; use human-readable time in dir names
For most users, "hostname" is the same for every crash (it is their hostname).

Not every problem even has "Application" field ("low on disk space" problem,
for example, doesn't), whereas any problem should have "Reason" field.

For non-root, it is useful to see which dumps are in /var/spool/abrt
(and aren't writable), and which are in $HOME/.abrt/spool.

Signed-off-by: Denys Vlasenko <dvlasenk@redhat.com>