don't use a status icon when the notification server supports persistence trac#127
Newer notifications servers (gnome-shell, or notification-daemon >= 0.6.0)
support persistence of notifications. This means that a bubble won't just
disappear after it is shown - it will stay until it is
seen/addressed/acted upon. This means we don't have to rely on a
status icon to provide the persistence for the message. In fact we
should remove the status icons since they will provide a confusing
duplication of this information