Better duplicate hash
Currently the duplicate hash of a crash for Bugzilla is computed from
the package NVR (name, version, release), path of the executable, and
the backtrace hash.

This is very unfortunate because the package NVR is changed frequently
(it contains Fedora OS version), and path of the executable is not a
good differentiator. When multiple executables from a single component
crash with similar backtrace, it is better to assume that they share
code and have single Bugzilla bug report.

This commit changes the code to compute the duplicate hash from the
component name and the backtrace hash.