run abrt oops log scanner as a separate service
In preparation for
https://fedorahosted.org/abrt/ticket/189
"abrtd should reload configuration when some of it's config files changes"
I would like to stop running abrt-dump-oops as a child of abrtd.

We experimented with it: we made abrt-ccpp hook (de)installation
a separate service, and in my opinion it ended up a better
solution than making it a part of daemon. Among other things,
it made abrt and abrt-addon-ccpp packages independent.

This change does the same for abrt-dump-oops.

Signed-off-by: Denys Vlasenko <dvlasenk@redhat.com>