wizard: hook selection checkboxes to reporters.
Checkboxes are controlled by the following elements in .xml:

<requires-items> ITEM1,ITEM2 </requires-items>
<exclude-items-by-default> ITEM1,ITEM2 </exclude-items-by-default>
<exclude-items-always> ITEM1,ITEM2 </exclude-items-always>
<exclude-binary-items> yes / no </exclude-binary-items>
<include-items-by-default> ITEM1,ITEM2 </include-items-by-default>

exclude-items-by-default and exclude-items-always can be "*"
meaning "all". include-items-by-default specifies which items
are included (checked) is exclude-items-by-default is "*".
Else, all are checked by default.

the set of iters NOT included is passed in $EXCLUDE_FROM_REPORT
variable. bugzilla, kerneloops, mailx, print and rhtsupport
are made aware of this variable.

Signed-off-by: Denys Vlasenko <dvlasenk@redhat.com>