rhbz#724838 - don't file kernel bugs if "tainted: B" is set
Kernel maintainers are not much happy that we filling bugs when kernel
is tainted.

Dave Jones said
---
If something has corrupted memory, we typically see a whole bunch of things
start crashing. This leads to situations where abrt files a lot of bugs, even
though most of them are just fallout from the first.

an example: bugs 710925 710930 710932 710933 710935 710942 710948 710955 710956
are all the same crash. The traces and addresses are different, but 'something
scribbled over memory' is the root cause here. Sadly out of all those bugs, we
actually somehow missed the first crash which is the most relevant one. (That
one would have not had the 'Tainted: G B' text)
---

First part of the patch is that we won't relay on /proc/sys/kernel/tainted and
we will parse the kernel bt manually, because of reading that file after an oops
is ALWAYS going to show it as tainted.

Second part of the patch is that whenever you want to prohibit reporting, put a
file with name *not-reportable* into <dump-dir>. That file could be empty or with
reasonable message why are you trying to prohibit reporting.

All 3 reporters plugins will be respect that file and it won't try to report anything.
Reporters will show only the messages from *not-reportable* and *reason*.

Signed-off-by: Nikola Pajkovsky <npajkovs@redhat.com>