Duplicate detection by coredump-level backtraces
Introduces gdb dependency to abrt-addon-ccpp.

Copypasted description of current dup detection algorithm:

This function is run after each post-create event is finished (there may
be multiple such events).

It first checks if there is CORE_BACKTRACE or UUID item in the dump
dir we are processing.

If there is a CORE_BACKTRACE, it iterates over all other dump
directories and computes similarity to their core backtraces (if
any).  If one of them is similar enough to be considered duplicate, the
function saves the path to the dump directory in question and returns 1
to indicate that we have indeed found a duplicate of currently processed
dump directory.  No more events are processed and program prints the
path to the other directory and returns failure.

If there is an UUID item (and no core backtrace), the function
again iterates over all other dump directories and compares this UUID to
their UUID. If there is a match, the path to the duplicate is saved and
1 is returned.

If duplicate is not found as described above, the function returns 0 and
we either process remaining events if there are any, or successfully
terminate processing of the current dump directory.