abrt-action-trim-files: delete up to 128 files between rescans, not one at a time
This fixes a performance problem I noticed while
working on another problem:
we scan debuginfo cache directory, find the "worst" file,
and if we need to trim the directory, we delete this one file
and repeat the scan. But scan is a long operation on large dirs.

This patch makes abrt-action-trim-files to delete more than one file
between rescans.

Signed-off-by: Denys Vlasenko <dvlasenk@redhat.com>