abrtd: allow parallel runs of "post-create" events
Before this patch, abrtd will be stuck waiting for "post-create to finish,
not accepting, for example, socket connections while it runs.
(On my system, "post-create" takes 1 minute because of broken sendmail).

This is the first part of the fix. It creates and hooks up a GIO
source made from "post-create" stdout to main loop, and consumes its output
concurrently with other tasks.

Signed-off-by: Denys Vlasenko <vda.linux@googlemail.com>