Integration with satyr
ABRT can now be built against satyr instead of btparser using
--with-satyr parameter to the configure script. RPMs are built against
btparser.

For the most part, the code is changed to use satyr and satyr-compat.h
is provided to wrap btparser in interface mimicking that of satyr. When
the migration to satyr is completed this file can be removed.

Note that the format of core_backtrace problem directory item is
incompatible between these two libraries.

Closes #590. Closes #591. Closes #592. Closes abrt/satyr#47.

Signed-off-by: Martin Milata <mmilata@redhat.com>
Signed-off-by: Jakub Filak <jfilak@redhat.com>