abrt-action-install-debuginfo: do not assume os.execvp never returns
I guess it's possible os.execvp() in Python doesn't return failure
but throws an exception. So it seems it might be a function which
never returns, and the code is ok as-is.
But my C gut feeling says "lets not assume that".
Thus, this change adds an error log and exit after os.execvp() call.

Signed-off-by: Denys Vlasenko <dvlasenk@redhat.com>