abrt-action-save-package-data: fix handling of ProcessUnpackaged on scripts
get_script_name() returns NULL if "python /path/to/something"
and /path/to/something isn't packaged, or __if it is blacklisted__.

But then the caller, SavePackageDescriptionToDebugDump,
returns 1 as intended... only if (!settings_bProcessUnpackaged).
This is wrong if /path/to/something is blacklisted -
in this case, it should be unconditional.

This change moves "if (is_path_blacklisted(executable))"
check out of get_script_name() into the caller.

While as it, eliminate double checking of !settings_bProcessUnpackaged
by also moving
 log("Interpreter crashed, but no packaged script detected: '%s'", cmdline);
tot he caller.

Signed-off-by: Denys Vlasenko <dvlasenk@redhat.com>
Signed-off-by: Jiri Moskovcak <jmoskovc@redhat.com>