abrt-handle-event: create DUMP_LOC/post-create.lock when running "post-create"
If lock can't be created fatally, such as DUMP_LOC doesn't exist,
or filesystem is RO, we won't run "post-create" at all.
Else, if there is another "post-create" holding it,
we wait for up to 100 seconds.

This change depends on create_symlink_lockfile() being exposed
in public API in libreport.

Signed-off-by: Denys Vlasenko <dvlasenk@redhat.com>
Signed-off-by: Martin Milata <mmilata@redhat.com>