abrt-cli list: use sha1 hash as short ids instead of @N thing. rhbz#906733
Once upon a time it was meant to allow addressing problems with @N
instead of full name. However, it is problematic - what the numbering
should be if e.g. --since DATE was used? Should we scan all directories
regardless because we want to show the same @N's as in full list?
Same problem with "--full List even reported problems" versus
!--full.

Instead, use git's trick of "shortcut IDs" based on SHA hashes.
In our case, use sha1 hash of directory name.

abrt-cli info STR will attempt to use STR as sha1 hash
of directory name if STR isn't an existing directory name,
is >= 5 chars long, and is a hex string.

Example:
    $ abrt-cli list
    ...
    @d9ca58e5b3e509759c582a5c162bfe24b23150dd
    Directory:      /var/tmp/abrt/oops-2013-08-27-14:08:13-9866-0
    time:           Tue 27 Aug 2013 02:08:13 PM CEST
    $ abrt-cli info d9ca5
    Directory:      /var/tmp/abrt/oops-2013-08-27-14:08:13-9866-0
    time:           Tue 27 Aug 2013 02:08:13 PM CEST

Signed-off-by: Denys Vlasenko <dvlasenk@redhat.com>
Signed-off-by: Jiri Moskovcak <jmoskovc@redhat.com>