fix debuginfo installer expecting user input from a pipe - closes #696
The debuginfo installer currently expects both build-ids and user input from a
pipe with analaze-core, when run from an UI.
We need to change it so that it passes the build-ids through a temporary file
and only uses stdin to communicate with the user.
Also makes the output file from analyze-core readable for others.

Signed-off-by: Petr Kubat <pkubat@redhat.com>
Signed-off-by: Jakub Filak <jfilak@redhat.com>