abrt-dump-oops: if backtrace is not seen, file an one-line oops. Fixes rhbz#812537
This allows to catch "Machine Check Exception:" oopses:
previously, they weren't saved because the code required
oops to have a backtrace to be saved, which MCEs lack.

Also, one of our "not_oops" testcases _is_ a kernel bug,
and with this change it is caught too.
Thus, not_oops3 test is renamed to oops7.

Added a test for MCE oops and tested it.

Signed-off-by: Denys Vlasenko <dvlasenk@redhat.com>