a-a-g-machine-id: suppress its failures in abrt_event.conf
If an event script exits with a non-zero code, abrtd deletes entire dump
directory.

Failures of a-a-g-machine-id must not lead to the deletion of a dump
directory.

It is also expected that dmidecode-less systems will print an error
message related to the fact that dmidecode doesn't work, hence
forwarding of STDOUT and STDERR to 'event_log' file.

Related to rhbz#1139552

Signed-off-by: Jakub Filak <jfilak@redhat.com>