gdb: make gdb aware of the abrt's debuginfo dir
A debuginfo package might ship an auto-loaded gdb script. If abrt
unpacks that package into the abrt's debuginfo cache dir and points gdb
to that directory, gdb refuses to auto-loaded that gdb scripts and
produces a plenty of warning messages that abrt writes to 'backtrace'
file.

The previous solution of this issue was to turn auto-load off completely
but it turned the pretty printer off too.

The correct solution is to add the abrt's debuginfo cache directory to
auto-load safe-path and auto-load scripts-dir settings.

Thanks Jan Kratochvil <jkratoch@redhat.com>

Requires: rhbz#1163339
Resolves: rhbz#1128637

Signed-off-by: Jakub Filak <jfilak@redhat.com>