add stuff necessary for Project Atomic
We are going to create a new package that will be installed in Atomic
hosts. The package has to pull the abrt core_dump helper without abrtd
and that's a bit problem because the StandaloneHook configuration option
must be set to 'yes' in '/etc/abrt/plugins/CCpp.conf' but we want to
provide an abrt configuration working out of the box. Hence adding
a new file CCpp_Atomic.conf that is a copy of CCpp.conf but tuned for
Atomic hosts. Another problem with missing abrtd is the requirement of
abrtd.service in abrt-ccpp.service, hence adding a new service file
abrt-coredump-helper.service without the requirement of abrtd.

Both new files will be shipped with the abrt-atomic package.

Signed-off-by: Jakub Filak <jfilak@redhat.com>