dbus: add new method to test existence of an element
It is sometimes necessary to check if some elemen exist, so this method
should be fast as much as it is possible to do this task over DBus.

I was thinking about calling the GetInfo method with a single element
but I refused this idea as it is inherently overcomplicated and error
prone.

Related: #1224984

Signed-off-by: Jakub Filak <jfilak@redhat.com>

Conflicts:
	doc/problems-service/org.freedesktop.Problems.xml.in