vmcore: generate 'reason' file in all cases
If kdump generates the dmesg log file (vmcore-dmesg.log), the vmcore's
post-create event doesn't generate 'reason' file. It is caused by
inappropriate use of 'abrt-dump-oops' where the event uses that helper
to parse the log file to generate 'backtrace' file instead of using it
with the '-u' argument to update the dump directory and create all
necessary files.

Signed-off-by: Jakub Filak <jfilak@redhat.com>