lib: don't expect kernel's version '2.6.*' or '3.*.*'
The function parsing kernel's version was expecting one of the strings
mentioned in Summary. Unfortunately, none of them appears in oopses for
kernel-4.*.*

I am not sure why the previous version didn't search for 'kernel-'
string, but I hope the previous authors know the reason. I can only
guess that 'kernel-' string is not always present, so I must not use it
in this commit. Hence, this commit switches to search by a regular
expression where I want to match a version string "\d.\d.\d" with
expectation of a release string in form of "-[^ )]+".

Signed-off-by: Jakub Filak <jfilak@redhat.com>