cli: adapt to PrivateReports
Restart the process as root if PrivateReports is enabled.

We want to run abrt-cli with root privileges if PrivateReports=yes by default,
because we think that this is the most comfortable work-flow for our users and
we are not sure know how to inform users about the need to run abrt-cli with
root privileges in a different way.

We want to fall back to user privileges, because if a user wants to report a
problem in RHEL-6, ABRT moves the corresponding dump directory to the user's
home, so the directory becomes inaccessible for root in some cases.

Signed-off-by: Jakub Filak <jfilak@redhat.com>

Conflicts:
	src/cli/abrt-cli-core.c