applet: fix problem info double free
There was a double free when an action function was called.
Problem info is freed when the notification is destroyed because the free
function is registred when the action is added
(notify_notification_add_action()). So the problem_info_unref() function in the
action function is moreover and causes the double free problem.

Related to rhbz#1211644

Signed-off-by: Matej Habrnal <mhabrnal@redhat.com>