cli: don't start reporting of not-reportable problems
If the reported problem data contains 'not-reportable' element, the
reporting process fails unexpectedly and after the reporter already spent some
time on it.

This commit ensures that the reporting process won't start, so
abrt-cli's behaviour will be consistent with ABRT GUI.

However, this is not an ideal solution because we might want to allow
the reporter to report the problem directly to developers via e-mail.

Closes #986

Signed-off-by: Jakub Filak <jfilak@redhat.com>