Rename uReport event
When I was adding the 'submit_uReport' event, I chose the prefix 'submit' in
order to avoid displaying the event in report-gtk | report-cli as a potential
report destination.

Now I need to rename the event to 'report_uReport' because we need to offer
it to users as a report destination.

Related: #1258474