CCpp: turn off compat cores
Disable compat cores because with systemd>=229, RLIMIT_CORE is
'unlimited' by default and RLIMIT_CORE!=0 was the signal to create a
core file in the old way. So, abrt-ccpp-hook creates the old way cores
everywhere on file systems by default and users don't know about it
because they did not enable it manually via 'ulimit -c unlimited'.

We could restore the old RLIMIT_CORE value by creating a systemd
configuration file with 'DefaultLimitCORE=0' but that would not be a
good solution for our users:
* you have to reboot, otherwise the configuration has no effect
* and therefore it is impossible to set the new value when
  abrt-ccpp.service is stopped and the users who expect
  RTLIMIT_CORE='unlimited' would be confused

We should probably adopt the behaviour of systemd-coredump. However,
that means that abrtd must not drop any problem it receives and should
mark those problems that would be dropped as NOT-REPORTABLE to prevent
users from filing bug reports for unfixable problems.

Find more details at:
https://github.com/systemd/systemd/issues/2643

Signed-off-by: Jakub Filak <jfilak@redhat.com>