Merge a-a-save-kernel-data with a-a-save-package-data
I forgot to do this while working on commit
650822d0d2438825515c0e41f06cd9fb77a18334

I do not think creating another tool doing almost the same thing was a
good idea. I believe we must keep relevant bits together and not to
spread functionality among myriads of tools just because someone do not
want to code in C.

I failed to find any justification for creating
'abrt-action-save-kernel-data' instead of teaching
'abrt-action-save-package-data' to do that job:
130dee46d601f3af6bb196d99a5c911335506adf

This commit changes contents of 'package' file from 'kernel' to
'kernel-$version' string. It should not break anything as I could not
find any justification for the 'kernel' string except the original
author's believe it should be like that (probably because of efficiency).

Just for the record, we do not need the version in 'package' file
because the version is already included in 'kernel' file.

This commit enables GPGCheck and ProcessUnpackaged options for
Kerneloopses and VMcores.

Signed-off-by: Jakub Filak <jfilak@redhat.com>